Fix torch stack error when there are multiple images with different sizes in the input
Previously, in `LlavaMetaModel.encode_images`, it will stack the features from different images in the input into a single tensor. However, when the sizes of image are different, it should stack but instead just return the list of the features. This is fine because in `BasicImageEncoder.forward` it will break down the features into a list anyway.